添加security mavenprofile

+++ item-bank-domain/src/main/java/com/mypro/ssm/po/rbac/UserRole.java
@@ -10,8 +10,8 @@ import java.io.Serializable;
  */
 public class UserRole implements Serializable {
 
-    private Integer userId;
-    private Integer roleId;
+    private Long userId;
+    private Long roleId;
 
     // Constructor
     public UserRole() {
@@ -20,26 +20,24 @@ public class UserRole implements Serializable {
     /**
      * full Constructor
      */
-    public UserRole(Integer userId, Integer roleId) {
+    public UserRole(Long userId, Long roleId) {
         this.userId = userId;
         this.roleId = roleId;
     }
 
-    // getter && setter
-    // 在setter方法最后加上"return this;"并把返回参数改为UserRole可以实现连缀设置属性
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public Integer getRoleId() {
+    public Long getRoleId() {
         return roleId;
     }
 
-    public void setRoleId(Integer roleId) {
+    public void setRoleId(Long roleId) {
         this.roleId = roleId;
     }
